llm-app/requirements.txt
pymongo
streamlit
elasticsearch
mistralai
stqdm